chore: re-organize all files
